integrations/demotest/buy_product_issue_1/requirements.txt
requests==2.24.0
python-dotenv==0.19.2